+++ src/MarginTrading.Core/IClientNotifyService.cs
@@ -4,6 +4,6 @@
     {
         void NotifyOrderChanged(Order order);
         void NotifyAccountChanged(IMarginTradingAccount account);
-        void NotifyAccountStopout(string clientId, string accountId, int positionsCount, double totalPnl);
+        void NotifyAccountStopout(string clientId, string accountId, int positionsCount, decimal totalPnl);
     }
 }
